fix: 통계포험여부 추가

+++ src/main/java/connectingstar/tars/habit/command/RunHabitCommandService.java
@@ -289,7 +289,13 @@ public class RunHabitCommandService {
         quitHabitRepository.save(quitHabit);
 
         habitAlertRepositoryCustom.deleteByRunHabitId(runHabit.getRunHabitId());
-        habitHistoryRepositoryCustom.deleteByRunHabitId(runHabit.getRunHabitId());
+        if(request.isVisibility()){
+            habitHistoryRepositoryCustom.updateByRunHabitId(runHabit.getRunHabitId());
+        }
+        else{
+            habitHistoryRepositoryCustom.deleteByRunHabitId(runHabit.getRunHabitId());
+        }
+
         runHabitRepository.delete(runHabit);
 
         return quitHabitMapper.toHabitDeleteResponse(quitHabit);
